refactor v2 trace map collection helpers
